Make API URL configurable
This is so that deployment is a bit easier on other systems.

Also found out the .gitignore file can be in a subdirectory and the base
file had wrong paths anyway. The actual effective one for solargraph was
in that subdir so I kept it and removed a lot from the base .gitignore.