Fixed bug refreshing current status
The refresh of current status was spamming the API. The issue was the
functions were calling 'fetch' in a client module which is not allowed
in next.js. The correct way in next.js is to call the 'use' method on a
promise that is passed in from the server module which sets up the
'fetch' call.

In addition, the 'fetch' call inside of the useEffect / setInterval
combo is not allowed in next.js. Instead, the client must invoke a
server function that calls 'fetch'.