Add daily chart click event
The daily chart now supports clicking on a data point to navigate to the
samples page with timestamps set for that day. This allows a user to
view the daily data and 'zoom in' on a day of interest.

A way to do this in react-ploty.js wasn't found although it appears that
the onClick event should be supported. Instead the Recharts library was
used and the plot refactored to use this more react friendly charting
library.

Part of this refactor consolidates the derived plot to use the same
component by passing in some extra metadata about the plot.