Pad sample data to visualise missing data
The sample data from the API simply omits missing samples. We want to
show missing samples on the line plot as blank spaces. To accomplish
this, samples with null value fields, ie. only timestamps, are added to
the plot data.

During this refactor, it was found that the data coming from the API is
in reverse time order so it is re-reversed in the UI. This was
previously thought to be a weird choice made in the recharts library so
the reversal was being done in the plotting component.